+++ rich/bar.py
@@ -79,7 +79,6 @@ class Bar:
                 remaining_bars -= 1
             if remaining_bars:
                 yield Segment(bar * remaining_bars, style)
-        yield Segment("\r")
 
     def __measure__(self, console: Console, max_width: int) -> Measurement:
         if self.width is not None:
@@ -97,6 +96,7 @@ if __name__ == "__main__":
     for n in range(0, 101, 1):
         bar.update_progress(n)
         console.print(bar)
+        console.file.write("\r")
         time.sleep(0.05)
     console.show_cursor(True)
     console.print()