+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2020-03-17
+
+### Added
+
+- CJK support
+- Console level highlight flag
+- Added encoding argument to Syntax.from_path
+
+### Changed
+
+- Dropped support for Windows command prompt (try https://www.microsoft.com/en-gb/p/windows-terminal-preview/)
+- Added task_id to Progress.track
+
 ## [0.7.2] - 2020-03-15
 
 ### Fixed